Check user when logging out
 - Implement Cancel button to cancel logout in alert

+++ src/screens/UserInfo/index.js
@@ -54,25 +54,29 @@ class UserInfo extends Component {
     }
   };
 
-  handleSignOutButton = async () => {
+  handleSignOutButton = () => {
     try {
       this.setState({ pending: true });
-      await AsyncStorage.removeItem('userInfo');
-      await AsyncStorage.removeItem('accessToken');
-      await AsyncStorage.removeItem('recentChallenge');
-      await SecureStore.deleteItemAsync('refreshToken');
-      await SecureStore.deleteItemAsync('keyChain');
-      this.setState({ user: null });
-      AlertIOS.alert('로그아웃 성공!', '보고싶을 거에요 🥺', [
+      Alert.alert('정말 로그아웃 하시겠어요?', '', [
         {
           text: 'OK',
-          onPress: () => this.goTo('Home'),
+          onPress: async () => {
+            await AsyncStorage.removeItem('userInfo');
+            await AsyncStorage.removeItem('accessToken');
+            await AsyncStorage.removeItem('recentChallenge');
+            await SecureStore.deleteItemAsync('refreshToken');
+            await SecureStore.deleteItemAsync('keyChain');
+            this.setState({ user: null, pending: false });
+            this.goTo('Home');
+          },
         },
-      ]);
+        {
+          text: 'Cancel',
+          onPress: () => this.setState({ pending: false }),
+        },
+      ], { cancelable: false });
     } catch (error) {
       AlertIOS.alert(error.message);
-    } finally {
-      this.setState({ pending: false });
     }
   };
